refactor(read-save-user-settings): narrow down UserSettingsRepository to getSetting & saveSetting methods returning preferencesDefaults if preferences null.
create getSetting within UserSettingsViewModel calling the repository's method. Retrieve settings from CountdownService to set the pomodoro duration & breaks.

+++ app/src/main/java/com/enriqueajin/pomidorki/presentation/UserSettingsEvent.kt
@@ -1,7 +1,7 @@
 package com.enriqueajin.pomidorki.presentation
 
 sealed interface UserSettingsEvent {
-    data class UpdatePomodoroDuration(val minutes: Int) : UserSettingsEvent
-    data class UpdateShortBreakDuration(val minutes: Int) : UserSettingsEvent
-    data class UpdateLongBreakDuration(val minutes: Int) : UserSettingsEvent
+    data class UpdatePomodoroDuration(val minutes: Long) : UserSettingsEvent
+    data class UpdateShortBreakDuration(val minutes: Long) : UserSettingsEvent
+    data class UpdateLongBreakDuration(val minutes: Long) : UserSettingsEvent
 }
